Arrangement in some classes
