+++ src/test/java/FeatureFiles/US_702_Mert.feature
@@ -1,23 +1,18 @@
 Feature: Login Functionality
 
-  Scenario: Login with null username and password
-
+  Background:
     Given Navigate to Magento Website
-    When Click on the Sign In button
+
+  Scenario: Login with null or wrong username and password
+    When Click on the sign In button
     Then Log in without filling in username and password
     And User should receive incorrect login warning
 
-    Then User should only fill in the e-mail section
+    And User should only fill in the e-mail section
     And User should receive password warning
 
-    Then User should only fill in the password section
+    And User should only fill in the password section
     And User should receive e-mail warning
 
-    Then User must enter wrong email and password
+    And User must enter wrong email and password
     And User should receive incorrect email and password warning
-
-    Then Finally, user must login to the account
-    And User should see successful login message
-
-    Then User should click on the Logout button
-    And User should be able to see the Forgot my Password button
